Refactor steam bridge
* split up in different files
* decrease complexity
